Split lockfile tests by subject
Separate manifest merging and path projection from lockfile regeneration so
each test module exercises one cohesive production API. Keep private support
code local and preserve the regeneration snapshot under its existing module.
