+++ tests/unit/test_bump_lockfiles.py
@@ -1,4 +1,4 @@
-"""Tests for lockfile regeneration after bump operations."""
+"""Tests for Cargo lockfile regeneration after bump operations."""
 
 from __future__ import annotations
 
@@ -7,13 +7,12 @@ import dataclasses as dc
 import operator
 import pathlib
 import shlex
-import string
 import tempfile
 import typing as typ
 from pathlib import Path
 
 import pytest
-from hypothesis import assume, given, settings
+from hypothesis import given, settings
 from hypothesis import strategies as st
 
 from lading.commands import bump_lockfiles
@@ -49,188 +48,6 @@ class _RecordingRunner:
         """Record one command invocation."""
         self.invocations.append(_Invocation(command=tuple(command), cwd=cwd))
         return self.result
-
-
-def _write_manifest(directory: Path) -> None:
-    """Create ``directory`` with a placeholder ``Cargo.toml`` inside it."""
-    directory.mkdir(parents=True, exist_ok=True)
-    (directory / "Cargo.toml").write_text(
-        '[package]\nname = "placeholder"\nversion = "0.1.0"\n',
-        encoding="utf-8",
-    )
-
-
-@st.composite
-def _manifest_merge_cases(
-    draw: st.DrawFn,
-) -> tuple[tuple[str, ...], tuple[str, ...]]:
-    """Generate configured manifests and arbitrarily ordered tracked lockfiles."""
-    directories = draw(
-        st.lists(
-            st.from_regex(r"[a-z][a-z0-9]{0,7}", fullmatch=True),
-            min_size=1,
-            max_size=8,
-            unique=True,
-        )
-    )
-    discovered_directories = draw(st.permutations(directories))
-    configured_directories = draw(st.lists(st.sampled_from(directories), unique=True))
-    configured_directories = draw(st.permutations(configured_directories))
-    dot_prefixes = draw(
-        st.lists(
-            st.booleans(),
-            min_size=len(configured_directories),
-            max_size=len(configured_directories),
-        )
-    )
-    configured = tuple(
-        f"{'./' if has_dot_prefix else ''}{directory}/Cargo.toml"
-        for directory, has_dot_prefix in zip(
-            configured_directories, dot_prefixes, strict=True
-        )
-    )
-    discovered = tuple(
-        f"{directory}/Cargo.lock" for directory in discovered_directories
-    )
-    return configured, discovered
-
-
-def test_merge_discovered_manifests_appends_tracked_manifests(
-    tmp_path: Path,
-) -> None:
-    """Discovered tracked-lockfile manifests follow configured entries, sorted."""
-    _write_manifest(tmp_path)
-    _write_manifest(tmp_path / "crates/ui")
-    _write_manifest(tmp_path / "fixtures/minimal")
-    runner = _RecordingRunner(
-        result=(
-            0,
-            "fixtures/minimal/Cargo.lock\nCargo.lock\ncrates/ui/Cargo.lock\n",
-            "",
-        )
-    )
-
-    merged = bump_lockfiles.merge_discovered_manifests(
-        tmp_path,
-        ("crates/ui/Cargo.toml",),
-        runner=runner,
-    )
-
-    assert merged == (
-        "crates/ui/Cargo.toml",
-        "Cargo.toml",
-        "fixtures/minimal/Cargo.toml",
-    ), f"unexpected merged manifests: {merged!r}"
-    assert runner.invocations == [
-        _Invocation(
-            command=("git", "ls-files", "**/Cargo.lock", "Cargo.lock"),
-            cwd=tmp_path,
-        )
-    ], f"unexpected git invocations: {runner.invocations!r}"
-
-
-def test_merge_discovered_manifests_deduplicates_configured_forms(
-    tmp_path: Path,
-) -> None:
-    """A manifest configured under an equivalent spelling is not re-appended."""
-    _write_manifest(tmp_path / "fixtures/minimal")
-    runner = _RecordingRunner(result=(0, "fixtures/minimal/Cargo.lock\n", ""))
-
-    merged = bump_lockfiles.merge_discovered_manifests(
-        tmp_path,
-        ("./fixtures/minimal/Cargo.toml",),
-        runner=runner,
-    )
-
-    assert merged == ("./fixtures/minimal/Cargo.toml",), (
-        f"equivalent configured manifest was re-appended: {merged!r}"
-    )
-
-
-def test_merge_discovered_manifests_returns_configured_outside_git(
-    tmp_path: Path,
-) -> None:
-    """A non-git workspace degrades to the configured manifests unchanged."""
-    runner = _RecordingRunner(
-        result=(
-            128,
-            "",
-            "fatal: not a git repository (or any of the parent directories): .git",
-        )
-    )
-
-    merged = bump_lockfiles.merge_discovered_manifests(
-        tmp_path,
-        ("crates/ui/Cargo.toml",),
-        runner=runner,
-    )
-
-    assert merged == ("crates/ui/Cargo.toml",), (
-        f"non-git fallback changed configured manifests: {merged!r}"
-    )
-
-
-@given(case=_manifest_merge_cases())
-@settings(max_examples=80, deadline=None)
-def test_merge_discovered_manifests_preserves_order_and_deduplicates(
-    case: tuple[tuple[str, ...], tuple[str, ...]],
-) -> None:
-    """Configured order wins while unique discoveries are appended sorted."""
-    configured, discovered = case
-    with tempfile.TemporaryDirectory() as temporary_directory:
-        workspace_root = Path(temporary_directory)
-        for lockfile in discovered:
-            _write_manifest(workspace_root / Path(lockfile).parent)
-        runner = _RecordingRunner(result=(0, "\n".join(discovered) + "\n", ""))
-
-        merged = bump_lockfiles.merge_discovered_manifests(
-            workspace_root,
-            configured,
-            runner=runner,
-        )
-
-        configured_paths = {
-            (workspace_root / manifest).resolve() for manifest in configured
-        }
-        expected_discovered = tuple(
-            sorted(
-                str(Path(lockfile).with_name("Cargo.toml"))
-                for lockfile in discovered
-                if (workspace_root / Path(lockfile).with_name("Cargo.toml")).resolve()
-                not in configured_paths
-            )
-        )
-        assert merged == configured + expected_discovered, (
-            f"merge did not preserve configured order and sorted discovery: {merged!r}"
-        )
-        assert len({
-            (workspace_root / manifest).resolve() for manifest in merged
-        }) == len(merged), f"merge retained equivalent manifest paths: {merged!r}"
-
-
-@given(
-    configured=st.lists(
-        st.from_regex(r"(?:\./)?[a-z][a-z0-9]{0,7}/Cargo\.toml", fullmatch=True),
-        max_size=8,
-    ).map(tuple)
-)
-@settings(max_examples=50, deadline=None)
-def test_merge_discovered_manifests_non_git_fallback_is_identity(
-    configured: tuple[str, ...],
-) -> None:
-    """Non-git discovery preserves every configured entry and its order."""
-    runner = _RecordingRunner(result=(128, "", "fatal: not a git repository"))
-    with tempfile.TemporaryDirectory() as temporary_directory:
-        merged = bump_lockfiles.merge_discovered_manifests(
-            Path(temporary_directory),
-            configured,
-            runner=runner,
-        )
-
-    assert merged == configured, f"non-git fallback changed input: {merged!r}"
-    assert len(runner.invocations) == 1, (
-        f"non-git discovery should invoke git once: {runner.invocations!r}"
-    )
 
 
 def test_regenerate_lockfiles_includes_workspace_manifest(tmp_path: Path) -> None:
@@ -315,39 +132,6 @@ def test_regenerate_lockfiles_deduplicates_root_manifest(tmp_path: Path) -> None
             str(tmp_path / "crates/nested/Cargo.toml"),
         ),
     ]
-
-
-def test_resolve_lockfile_paths_reports_dry_run_targets(tmp_path: Path) -> None:
-    """Dry-run reporting can resolve lockfiles without invoking Cargo."""
-    lockfiles = bump_lockfiles.resolve_lockfile_paths(
-        tmp_path,
-        ("Cargo.toml", "crates/nested/Cargo.toml"),
-    )
-
-    assert lockfiles == (
-        tmp_path / "Cargo.lock",
-        tmp_path / "crates/nested/Cargo.lock",
-    )
-
-
-@pytest.mark.parametrize(
-    ("manifest", "expected_message"),
-    [
-        ("../outside/Cargo.toml", "within the workspace"),
-        ("Cargo.lock", "Cargo.toml file"),
-        ("crates/nested/foo.toml", "Cargo.toml file"),
-    ],
-)
-def test_resolve_lockfile_paths_rejects_invalid_targets(
-    tmp_path: Path,
-    manifest: str,
-    expected_message: str,
-) -> None:
-    """Configured manifests must stay in-workspace and name Cargo.toml."""
-    with pytest.raises(
-        bump_lockfiles.LockfileRegenerationError, match=expected_message
-    ):
-        bump_lockfiles.resolve_lockfile_paths(tmp_path, (manifest,))
 
 
 @pytest.mark.parametrize(
@@ -460,147 +244,6 @@ def test_regenerate_lockfiles_wraps_runner_exceptions(tmp_path: Path) -> None:
 
 # ---------------------------------------------------------------------------
 # Hypothesis property tests for manifest-path resolution (issue #93)
-# ---------------------------------------------------------------------------
-
-_SEGMENT = st.text(
-    alphabet=string.ascii_lowercase + string.digits + "_-",
-    min_size=1,
-    max_size=10,
-)
-
-
-@st.composite
-def _inside_dir(draw: st.DrawFn) -> list[str]:
-    """Draw a relative directory path that stays within the workspace.
-
-    Alongside real segments and redundant ``.`` segments, this emits safe
-    ``..`` parent-traversal segments that never ascend above the workspace
-    root after normalisation: a ``..`` is only produced while a prior real
-    segment remains to cancel it (the running segment balance never drops
-    below zero). Cases such as ``crate/../Cargo.toml`` are therefore
-    exercised without allowing traversal above the root.
-    """
-    length = draw(st.integers(min_value=0, max_value=4))
-    components: list[str] = []
-    depth = 0
-    for _ in range(length):
-        options = [_SEGMENT, st.just(".")]
-        if depth > 0:
-            options.append(st.just(".."))
-        segment = draw(st.one_of(*options))
-        if segment == "..":
-            depth -= 1
-        elif segment != ".":
-            depth += 1
-        components.append(segment)
-    return components
-
-
-# Relative directory paths that stay inside the workspace, optionally with
-# redundant "." segments and safe ".." traversals which normalise away.
-_INSIDE_DIR = _inside_dir()
-
-
-def _manifest_string(components: list[str]) -> str:
-    """Render a manifest path string from directory ``components``."""
-    return "/".join((*components, "Cargo.toml"))
-
-
-@given(dirs=st.lists(_INSIDE_DIR, max_size=6))
-@settings(max_examples=60, deadline=None)
-def test_inside_manifests_resolve_to_sibling_lockfiles(
-    dirs: list[list[str]],
-) -> None:
-    """Any in-workspace manifest string yields a sibling Cargo.lock path.
-
-    Also pins the ordering and deduplication invariants: the workspace root
-    lockfile is always first, and resolved paths are unique.
-    """
-    with tempfile.TemporaryDirectory(prefix="lading-bump-lockfiles-") as tmp:
-        workspace_root = Path(tmp)
-        manifests = [_manifest_string(components) for components in dirs]
-
-        lockfiles = bump_lockfiles.resolve_lockfile_paths(workspace_root, manifests)
-
-        resolved_root = workspace_root.resolve()
-        assert lockfiles[0] == resolved_root / "Cargo.lock", (
-            "workspace root Cargo.lock must be the first resolved lockfile"
-        )
-        assert len(set(lockfiles)) == len(lockfiles), (
-            "resolved lockfiles must be unique"
-        )
-        for lockfile_path in lockfiles:
-            assert lockfile_path.name == "Cargo.lock", (
-                f"resolved path must be a sibling Cargo.lock: {lockfile_path}"
-            )
-            assert lockfile_path.parent == lockfile_path.parent.resolve(), (
-                f"lockfile parent must be a normalised path: {lockfile_path}"
-            )
-            assert lockfile_path.is_relative_to(resolved_root), (
-                f"lockfile must stay within the workspace root: {lockfile_path}"
-            )
-        # Build the expected ordered tuple: the workspace root Cargo.lock first,
-        # then the sibling lockfile for each manifest in execution order, with
-        # duplicates removed while preserving that order.
-        expected = tuple(
-            dict.fromkeys(
-                [resolved_root / "Cargo.lock"]
-                + [
-                    workspace_root.joinpath(*components, "Cargo.toml").resolve().parent
-                    / "Cargo.lock"
-                    for components in dirs
-                ]
-            )
-        )
-        assert lockfiles == expected, (
-            "resolved lockfiles must preserve manifest execution order "
-            "without duplicates"
-        )
-
-
-@given(spellings=st.lists(st.sampled_from(["Cargo.toml", "./Cargo.toml"]), max_size=4))
-@settings(max_examples=20, deadline=None)
-def test_root_manifest_spellings_deduplicate_to_one_invocation(
-    spellings: list[str],
-) -> None:
-    """Every spelling of the root manifest produces exactly one root entry."""
-    with tempfile.TemporaryDirectory(prefix="lading-bump-lockfiles-") as tmp:
-        workspace_root = Path(tmp)
-
-        lockfiles = bump_lockfiles.resolve_lockfile_paths(workspace_root, spellings)
-
-        assert lockfiles == (workspace_root.resolve() / "Cargo.lock",)
-
-
-@given(
-    escape_depth=st.integers(min_value=1, max_value=3),
-    suffix=st.lists(_SEGMENT, max_size=2),
-)
-@settings(max_examples=30, deadline=None)
-def test_escaping_manifests_are_rejected(
-    escape_depth: int,
-    suffix: list[str],
-) -> None:
-    """Any manifest path escaping the workspace root raises an error."""
-    with tempfile.TemporaryDirectory(prefix="lading-bump-lockfiles-") as tmp:
-        # Anchor inside a subdirectory so ".." segments cannot accidentally
-        # resolve back inside the temporary root.
-        workspace_root = Path(tmp) / "workspace"
-        workspace_root.mkdir()
-        # A suffix re-entering the workspace directory would resolve back
-        # inside and legitimately pass validation; exclude that case.
-        assume(suffix[:1] != ["workspace"])
-        escaping = "/".join(([".."] * escape_depth) + [*suffix, "Cargo.toml"])
-
-        with pytest.raises(
-            bump_lockfiles.LockfileRegenerationError,
-            match="must stay within the workspace",
-        ):
-            bump_lockfiles.resolve_lockfile_paths(workspace_root, (escaping,))
-
-
-# ---------------------------------------------------------------------------
-# Aggregated failure handling (issue #84)
 # ---------------------------------------------------------------------------
